iommu/samsung: drop useless cast (__ffs()) (v6.10 follow-up)
Upstream commit [1] changed the __ffs() return type from unsigned long
to unsigned int.

We can drop the useless cast here, the compiler will do the right thing
for either version.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=9c313ccdfc079f71f16c9b3d3a6c6d60996b9367 [1]
      ("bitops: Change function return types from long to int")
Test: TH
Bug: 355381571
(cherry picked from commit 438436bc5388fe3910bf9635d52ef0adba9862d3)

+++ drivers/iommu/samsung-iommu-fault.c
@@ -446,9 +446,9 @@ static void sysmmu_show_fault_information(struct sysmmu_drvdata *drvdata,
 static void sysmmu_get_interrupt_info(struct sysmmu_drvdata *data, unsigned int *intr_type,
 				      unsigned int *vid, sysmmu_iova_t *addr, bool is_secure)
 {
-	u32 istatus;
+	unsigned int istatus;
 
-	istatus = (unsigned int)__ffs(__sysmmu_get_intr_status(data, is_secure));
+	istatus = __ffs(__sysmmu_get_intr_status(data, is_secure));
 	*vid = istatus / 4;
 	*intr_type = istatus % 4;
 	*addr = __sysmmu_get_fault_address(data, *vid, is_secure);
